% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_boundaries.R
\name{admin_boundaries}
\alias{admin_boundaries}
\title{Administrative boundaries}
\usage{
admin_boundaries(type)
}
\arguments{
\item{type}{one of}
}
\value{
data frame
}
\description{
This is a function returning admin area boundaries
}
\examples{

plot(admin_boundaries("state"), max.plot = 1)


}
