% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_boundaries.R
\name{areas}
\alias{areas}
\title{Administrative boundaries}
\source{
\url{https://www.ehealthafrica.org/}
}
\usage{
areas(type)
}
\arguments{
\item{type}{one of "ward", "lga" and "state"}
}
\value{
data frame in \code{sf} package format
}
\description{
Function returning local administrative area boundaries in {sf} package format.
}
\details{
Recognized administrative areas are:
\itemize{
  \item{"state"} {States of Nigeria.}
  \item{"lga"} {Local Government Areas of Nigeria.}
  \item{"ward"} {Wards of Nigeria.}
}
}
\examples{
states <- areas("state")
plot(sf::st_geometry(state))

}
