% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_nigeriadata}
\alias{download_nigeriadata}
\title{Download Data on Nigeria}
\usage{
download_nigeriadata(item)
}
\arguments{
\item{item}{A character vector representing the kind of data to be downloaded.}
}
\description{
Download Data on Nigeria
}
