% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.R
\name{downloader}
\alias{downloader}
\title{Downloader}
\usage{
downloader(file)
}
\arguments{
\item{file}{file to be downloaded (or not...)}
}
\description{
Internal function - generic downloader, used to serve the data files
}
\keyword{internal}
