\name{farmlands}
\alias{farmlands}
\title{
Farmland Datasets in Nigeria
}
\description{
This function loads the datasets of all farmlands in NIgeria
}
\usage{
farmlands()
}

\references{
The Data and the geoserver link is provided by eHealth Systems Africa - https://www.ehealthafrica.org/ and also available on their data portal - https://data.ehealthafrica.org/
}
\author{
Joseph Oladokun
}

\examples{
farmlands_data <- farmlands()
}
