\name{industrial_sites}
\alias{industrial_sites}
\title{
Dataset for Industrial Sites in Nigeria
}
\description{
This function returns the datasets containing information on industrial sites in Nigeria
}
\usage{
industrial_sites()
}

\references{
The Data and the geoserver link is provided by eHealth Systems Africa - https://www.ehealthafrica.org/ and also available on their data portal - https://data.ehealthafrica.org/
}
\author{
Joseph Oladokun
}

\examples{
industrial_sites_data <- industrial_sites()
}
