\name{lga_admin_boundaries}
\alias{lga_admin_boundaries}
\title{
Dataset for adminstrative boundaries for Local Government Areas in Nigeria
}
\description{
This function returns the datasets of administrative boundaries in Nigeria at the local government areas level
}
\usage{
lga_admin_boundaries()
}

\references{
The Data and the geoserver link is provided by eHealth Systems Africa - https://www.ehealthafrica.org/ and also available on their data portal - https://data.ehealthafrica.org/
}
\author{
Joseph Oladokun
}

\examples{
lga_boundaries_data <- lga_admin_boundaries()
}
