\name{markets}
\alias{markets}
\title{
Dataset for Markets in Nigeria
}
\description{
This function returns the datasets containing geospatial and non-spatial information of Markets in Nigeria
}
\usage{
markets()
}

\references{
The Data and the geoserver link is provided by eHealth Systems Africa - https://www.ehealthafrica.org/ and also available on their data portal - https://data.ehealthafrica.org/
}
\author{
Joseph Oladokun
}

\examples{
market_data <- markets()
}
