% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nigeriadata.R
\docType{package}
\name{nigeriadata-package}
\alias{nigeriadata-package}
\title{This is nigeriadata package; it will be super awesome, once it is done:)}
\description{
This is nigeriadata package; it will be super awesome, once it is done:)
}
