\name{petrol_stations}
\alias{petrol_stations}
\title{
Dataset for Petrol/Filling/Gas Stations in Nigeria
}
\description{
This function returns the datasets containing information on Petrol/Filling/Gas Station in Nigeria
}
\usage{
petrol_stations()
}

\references{
The Data and the geoserver link is provided by eHealth Systems Africa - https://www.ehealthafrica.org/ and also available on their data portal - https://data.ehealthafrica.org/
}
\author{
Joseph Oladokun
}

\examples{
petrol_stations_data <- petrol_stations()
}
