% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_of_interest.R
\name{poi}
\alias{poi}
\title{Points of Interest}
\source{
\url{https://www.ehealthafrica.org/}
}
\usage{
poi(type)
}
\arguments{
\item{type}{one of "police" and "fuel"}
}
\value{
data frame in \code{sf} package format
}
\description{
Function returning POIs in {sf} package format.
}
\details{
Recognized points of interest are:
\itemize{
  \item{"police"} {police stations}
  \item{"fuel"} {fuel stations}
}
}
\examples{
plot(st_geometry(areas("state")))
plot(st_geometry(poi("police")), add = T, pch = 4, col = "red")

}
