\name{police_stations}
\alias{police_stations}
\title{
Dataset for Police Stations in Nigeria
}
\description{
This function returns the datasets containing geospatial and non-spatial information about Police Stations in Nigeria
}
\usage{
police_stations()
}

\references{
The Data and the geoserver link is provided by eHealth Systems Africa - https://www.ehealthafrica.org/ and also available on their data portal - https://data.ehealthafrica.org/
}
\author{
Joseph Oladokun
}

\examples{
police_station_data <- police_stations()
}
