\name{prisons}
\alias{prisons}
\title{
Dataset for Prisons in Nigeria
}
\description{
This function returns the datasets containing geospatial and non-spatial information of Prisons in Nigeria
}
\usage{
prisons()
}

\references{
The Data and the geoserver link is provided by eHealth Systems Africa - https://www.ehealthafrica.org/ and also available on their data portal - https://data.ehealthafrica.org/
}
\author{
Joseph Oladokun
}

\examples{
prison_data <- prisons()
}
