\name{schools}
\alias{schools}
\title{
Dataset for Schools in Nigeria
}
\description{
This function returns the datasets containing geospatial and non-spatial information of Schools in Nigeria
}
\usage{
schools()
}

\references{
The Data and the geoserver link is provided by eHealth Systems Africa - https://www.ehealthafrica.org/ and also available on their data portal - https://data.ehealthafrica.org/
}
\author{
Joseph Oladokun
}

\examples{
school_data <- schools()
}
