\name{state_admin_boundaries}
\alias{state_admin_boundaries}
\title{
Dataset for adminstrative boundaries for States in Nigeria
}
\description{
This function returns the datasets of administrative boundaries in Nigeria at the State level
}
\usage{
state_admin_boundaries()
}

\references{
The Data and the geoserver link is provided by eHealth Systems Africa - https://www.ehealthafrica.org/ and also available on their data portal - https://data.ehealthafrica.org/
}
\author{
Joseph Oladokun
}

\examples{
state_boundaries_data <- state_admin_boundaries()
}
