\name{ward_admin_boundaries}
\alias{ward_admin_boundaries}
\title{
Dataset for adminstrative boundaries for Wards in Nigeria
}
\description{
This function returns the datasets of administrative boundaries in Nigeria at ward level
}
\usage{
ward_admin_boundaries()
}

\references{
The Data and the geoserver link is provided by eHealth Systems Africa - https://www.ehealthafrica.org/ and also available on their data portal - https://data.ehealthafrica.org/
}
\author{
Joseph Oladokun
}

\examples{
ward_boundaries_data <- ward_admin_boundaries()
}
