% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_boundaries.R
\name{areas}
\alias{areas}
\title{Administrative boundaries}
\usage{
areas(type)
}
\arguments{
\item{type}{one of}
}
\value{
data frame
}
\description{
This is a function returning admin area boundaries
}
\examples{

states <- areas("state")
plot(sf::st_geometry(states))


}
