% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nigeriadata.R
\docType{package}
\name{nigeriadata-package}
\alias{nigeriadata-package}
\title{This is nigeriadata package; it will be super awesome, once it is done:)}
\description{
At present it is more of a skeleton; only two functions are implemented:
\itemize{
  \item{areas()} {returns administrative areas}
  \item{poi()} {returns points of interest}
}
}
\details{
All objects are returned in format of \code{sf} package.
}
